chore: add tests for deleting reservation

+++ server/internal/domain/entity/reservation.go
@@ -7,16 +7,16 @@ import (
 )
 
 type Reservation struct {
-	ID         int64           `json:"id"`
+	ID         int             `json:"id"`
 	ExternalID *string         `json:"external_id"`
 	UserID     string          `json:"user_id"`
 	CampusType enum.CampusType `json:"campus_type"`
 	RoomID     string          `json:"room_id"`
 	Date       time.Time       `json:"date"`
-	FromHour   int32           `json:"from_hour"`
-	FromMinute int32           `json:"from_minute"`
-	ToHour     int32           `json:"to_hour"`
-	ToMinute   int32           `json:"to_minute"`
+	FromHour   int             `json:"from_hour"`
+	FromMinute int             `json:"from_minute"`
+	ToHour     int             `json:"to_hour"`
+	ToMinute   int             `json:"to_minute"`
 	BookerName *string         `json:"booker_name"`
 	CreatedAt  time.Time       `json:"created_at"`
 }
